+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# unreleased
+# 0.3.2
 
 - Add label content block to edit component
 - Introduce `remove_content!` and make its non-bang pendant tolerate missing blocks